refactor(misc): extract regulations, banner, announcement, development_environment, object [Phase 1,3,4,5,6,8,9]
