+++ dojo/announcement/__init__.py
@@ -1,0 +1,1 @@
+import dojo.announcement.admin  # noqa: F401
