+++ dojo/announcement/api/__init__.py
@@ -1,0 +1,1 @@
+path = "announcements"  # noqa: RUF067
