+++ dojo/banner/__init__.py
@@ -1,0 +1,1 @@
+import dojo.banner.admin  # noqa: F401
