+++ dojo/banner/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+
+from dojo.banner.models import BannerConf
+
+admin.site.register(BannerConf)
