+++ dojo/development_environment/__init__.py
@@ -1,0 +1,1 @@
+import dojo.development_environment.admin  # noqa: F401
