+++ dojo/development_environment/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+
+from dojo.development_environment.models import Development_Environment
+
+admin.site.register(Development_Environment)
