+++ dojo/development_environment/api/__init__.py
@@ -1,0 +1,1 @@
+path = "development_environments"  # noqa: RUF067
