+++ dojo/development_environment/api/serializer.py
@@ -1,0 +1,9 @@
+from rest_framework import serializers
+
+from dojo.development_environment.models import Development_Environment
+
+
+class DevelopmentEnvironmentSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Development_Environment
+        fields = "__all__"
