+++ dojo/development_environment/ui/forms.py
@@ -1,0 +1,15 @@
+from django import forms
+
+from dojo.development_environment.models import Development_Environment
+
+
+class Development_EnvironmentForm(forms.ModelForm):
+    class Meta:
+        model = Development_Environment
+        fields = ["name"]
+
+
+class Delete_Dev_EnvironmentForm(forms.ModelForm):
+    class Meta:
+        model = Development_Environment
+        fields = ["id"]
