+++ dojo/object/__init__.py
@@ -1,0 +1,1 @@
+import dojo.object.admin  # noqa: F401
