+++ dojo/regulations/__init__.py
@@ -1,0 +1,1 @@
+import dojo.regulations.admin  # noqa: F401
