+++ dojo/regulations/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+
+from dojo.regulations.models import Regulation
+
+admin.site.register(Regulation)
