+++ dojo/regulations/api/__init__.py
@@ -1,0 +1,1 @@
+path = "regulations"  # noqa: RUF067
