+++ dojo/regulations/api/urls.py
@@ -1,0 +1,7 @@
+from dojo.regulations.api import path
+from dojo.regulations.api.views import RegulationsViewSet
+
+
+def add_regulations_urls(router):
+    router.register(path, RegulationsViewSet, basename="regulations")
+    return router
