+++ dojo/regulations/ui/views.py
@@ -8,8 +8,8 @@ from django.shortcuts import render
 from django.urls import reverse
 
 from dojo.authorization.authorization import user_has_configuration_permission_or_403
-from dojo.forms import RegulationForm
-from dojo.models import Regulation
+from dojo.regulations.models import Regulation
+from dojo.regulations.ui.forms import RegulationForm
 from dojo.utils import add_breadcrumb
 
 logger = logging.getLogger(__name__)
